requirements.txt
langgraph>=0.2.30
langchain>=0.2.10
langchain-openai>=0.1.22
pydantic>=2.7.0
python-dotenv>=1.0.1
langchain-community>=0.3.7
pypdf>=4.2.0
chromadb>=0.5.18
pymupdf4llm>=0.0.17
pdfplumber>=0.11.4
streamlit>=1.28.0
audio-recorder-streamlit>=0.0.8
streamlit-webrtc>=0.47.0
requests>=2.31.0
# 本地ASR依赖
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
edge-tts>=6.1.9
numpy>=1.24.0
torch>=2.0.0
